requirements.txt
Django==1.10
django-filter==0.14.0
djangorestframework==3.4.3
djangorestframework-bulk==0.2.1
-e git://github.com/danxshap/django-inline-ordering.git@908d9c9f915e7ecf61226bff535afa3bb6657a6d#egg=django-inline-ordering
